+++ CHANGELOG.md
@@ -12,6 +12,8 @@
 - [x] Validation
 - [x] Image hover preview
 - [x] Peek definition for classes and IDs
+- [x] Go to definition on classes and IDs
+- [x] Find all references for classes and IDs
 
 Features that have a checkmark are complete and available for
 download in the